Placed prod and test ennvironment on different server

+++ deploy.sh
@@ -19,10 +19,22 @@ mkdir -p ./build
 
 # copy the built scheduler to a location easy to zip
 rsync -av --progress scheduler ./build --exclude node_modules --exclude .yarn
+
+if [[ ! -f ./build/scheduler/.env.production ]] ; then
+    echo 'File "env.production" for scheduler is not there, aborting.'
+    exit
+fi
+
 cp ./build/scheduler/.env.production ./build/scheduler/.env
 
 # copy the built web api to a location easy to zip
 rsync -av --progress webapi ./build --exclude node_modules --exclude .yarn
+
+if [[ ! -f ./build/webapi/.env.production ]] ; then
+    echo 'File "env.production" for webapi is not there, aborting.'
+    exit
+fi
+
 cp ./build/webapi/.env.production ./build/webapi/.env
 
 # copy the built website to a location easy to zip
@@ -33,6 +45,11 @@ rsync -av --progress backoffice/public ./build/website
 
 rm -f ./build.zip
 zip -q -r --symlinks ./build.zip ./build
+
+if [[ ! -f ./docker/environments/prod/.env ]] ; then
+    echo 'File ".env" for docker environment is not there, aborting.'
+    exit
+fi
 
 # create docker folder
 ssh root@45.91.168.78 "mkdir -p /home/symbro/docker/containers;mkdir -p /home/symbro/docker/environments/prod;"
